+++ Application/MainCpp/inc/commands.hpp
@@ -1,0 +1,19 @@
+#ifndef MAINCPP_INC_COMMANDS_HPP_
+#define MAINCPP_INC_COMMANDS_HPP_
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void init_commands();
+void tick_commands();
+
+#ifdef __cplusplus
+}
+#endif
+
+
+
+
+
+#endif /* MAINCPP_INC_COMMANDS_HPP_ */
