+++ Application/MainCpp/inc/measurements.hpp
@@ -1,0 +1,16 @@
+#ifndef MAINCPP_INC_MEASUREMENTS_HPP_
+#define MAINCPP_INC_MEASUREMENTS_HPP_
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void init_measurements();
+void tick_measurements();
+
+#ifdef __cplusplus
+}
+#endif
+
+
+#endif /* MAINCPP_INC_MEASUREMENTS_HPP_ */
