+++ Core/Src/crc.cpp
@@ -1,4 +1,4 @@
-#include "crc.h"
+#include <crc.hpp>
 
 //CRC
 #define POLYNOMIAL 0x131 //P(x)=x^8+x^5+x^4+1 = 100110001
