+++ Core/Src/main.c
@@ -22,7 +22,6 @@
 
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
-#include "crc.h"
 #include <stdio.h>
 #include <string.h>
 #include "main_cpp.hpp"
@@ -58,52 +57,6 @@ UART_HandleTypeDef huart3;
 DMA_HandleTypeDef hdma_usart3_tx;
 
 /* USER CODE BEGIN PV */
-#define UART_BUF_LEN 255
-unsigned char UART3_rxBuffer[UART_BUF_LEN];
-
-#define ADC_BUF_LEN 4
-uint32_t adc_buf[ADC_BUF_LEN];
-
-#define ADC_IN1_PA0_POUT     0
-#define ADC_IN11_PB0_PROX    1
-#define ADC_IN6_PC0_I_MOT    2
-#define ADC_IN7_PC1_S_MOT    3
-
-#define SFM3219_ADDRESS     0x2E
-#define SFM3219_START_AIR   0x3608
-
-typedef enum{
-    I2C_INIT = 0,
-    I2C_READ = 1
-}i2c_states;
-
-int32_t offsetPout = -366;
-int32_t offsetPprox = -366;
-
-#define MEASURES_BUF_LEN 5
-int32_t Measures[MEASURES_BUF_LEN];
-#define MEAS_QOUT     0
-#define MEAS_POUT     1
-#define MEAS_PPROX    2
-#define MEAS_I_MOT    3
-#define MEAS_S_MOT    4
-
-char const* data_names[] =
-{
-        "time",
-        "qout",
-        "pout",
-        "pprox",
-        "mot_speed",
-        "mot_current"
-};
-
-uint32_t target_motor_speed     = 20000;    // 20000 rpm
-uint32_t target_motor_qout      = 1000;     // 10 L/min
-
-uint32_t cmd_motor              = 0; //100;  // 10%
-uint32_t cmd_peep               = 0; //200;  // 20%
-uint32_t cmd_valve              = 0;    // 0%
 
 /* USER CODE END PV */
 
@@ -124,247 +77,10 @@ static void MX_TIM2_Init(void);
 /* Private user code ---------------------------------------------------------*/
 /* USER CODE BEGIN 0 */
 
-void printFloats(float in[], int size)
-{
-    char buffer[80]="";
-    for (int i = 0; i < size; i++)
-    {
-        char txt[10];
-        sprintf(txt, "%.2f ", in[i]);
-        strcat(buffer, txt);
-    }
-    strcat(buffer, "\n");
-    HAL_UART_Transmit_DMA(&huart3, (uint8_t*)buffer, (uint8_t)(strlen(buffer)));
-}
-
-
-void printFloatsTelePlot(float in[], char const* names[], int size)
-{
-    char buffer[256]="";
-    for (int i = 0; i < size; i++)
-    {
-        char txt[64];
-        sprintf(txt, ">%s:%.2f ", names[i], in[i]);
-        strcat(buffer, txt);
-        strcat(buffer, "\r\n");
-    }
-    strcat(buffer, "\0");
-    HAL_UART_Transmit_DMA(&huart3, (uint8_t*)buffer, (uint8_t)(strlen(buffer)));
-}
-
-const int32_t MAX_CURRENT   = 300;
-const int32_t MAX_SPEED     = 60000;
-const int32_t MAX_RAW_ADC   = 4095;
-const int32_t MAX_RAW_I2C   = 0xFFFF;
-
-const int32_t PRESSURE_OFFSET   = 300;
-const int32_t PRESSURE_GAIN     = 60000;
-
-int32_t VoltageTo01mbar(int32_t volt, int32_t offset)
-{
-    return ( ( ( ( volt - 33 ) * 13790 ) / 264 ) - 6895) - offset;
-}
-
-int32_t RawADCToVolt(int32_t raw)
-{
-    return raw * 330 / MAX_RAW_ADC;
-}
-
-int32_t RawSFM3019ToLmin(int32_t raw, int32_t offset)
-{
-    return ( ( 100 * ( raw + 24576 ) ) / 170 ) - offset;
-}
-
-int32_t RawToCal(int32_t raw, int32_t max_cal, int32_t max_raw)
-{
-    return (raw * max_cal) / max_raw;
-}
-
-void UpdatePWM1(uint32_t per1000)
-{
-    TIM15->CCR1 = (per1000 * TIM15_PERIOD_TICKS) /1000;
-}
-
-void UpdatePWM2(uint32_t per1000)
-{
-    TIM15->CCR2 = (per1000 * TIM15_PERIOD_TICKS) /1000;
-}
-
-void UpdatePWM3(uint32_t per1000)
-{
-    TIM2->CCR1 = (per1000 * TIM2_PERIOD_TICKS) /1000;
-}
-
-int CMD_MOTOR_INSPI     = 2500;
-int CMD_MOTOR_EXPI      = 500;
-
-int INSPI_TIME          = 3000; //1000;
-int EXPI_TIME           = 3000; //2000;
-
-
-uint8_t cmd[3] = {0x36, 0x08, 0x00};
-uint8_t crc = 0;
-uint8_t i2c_state = 0;
-uint8_t i2c_retry = 0;
-
-int32_t rawQout = 0;
-int32_t offsetQout = -57;
-
-void InitQoutSensor()
-{
-    crc = SF04_CalcCrc (cmd, 2);
-    cmd[2] = crc;
-
-    HAL_GPIO_WritePin(CmdQout_GPIO_Port, CmdQout_Pin, GPIO_PIN_SET);
-    i2c_state = I2C_INIT;
-}
-
-#define I2C_INIT_ERROR_MSG "INIT_ERROR\n\0"
-#define I2C_CRC_ERROR_MSG  "CHECKSUM_ERROR\n\0"
-#define I2C_READ_ERROR_MSG "READING_ERROR\n\0"
-
-void ReadQoutSensor()
-{
-    static int tick_i2c = 0;
-    rawQout = -24576;
-
-    if (i2c_state == I2C_INIT)
-    {
-        if (tick_i2c <= 300)
-        {
-            HAL_GPIO_WritePin(CmdQout_GPIO_Port, CmdQout_Pin, GPIO_PIN_SET);
-        }
-        else if (tick_i2c <= 400)
-        {
-            HAL_GPIO_WritePin(CmdQout_GPIO_Port, CmdQout_Pin, GPIO_PIN_RESET);
-        }
-        else
-        {
-            uint8_t status = HAL_I2C_Master_Transmit(&hi2c1, SFM3219_ADDRESS<<1, cmd, 3, 1000);
-            if (status == HAL_OK)
-            {
-                i2c_state = I2C_READ;
-                i2c_retry = 0;
-            }
-            else
-            {
-                char buffer[] = I2C_INIT_ERROR_MSG;
-                HAL_UART_Transmit(&huart3, (uint8_t*)buffer, (uint8_t)(strlen(buffer)), 100U);
-            }
-            tick_i2c = 0;
-        }
-        tick_i2c++;
-    }
-    else if (i2c_state == I2C_READ)
-    {
-        uint8_t i2c_rcv_buff[3] = {0x00, 0x00, 0x00};
-        uint8_t status = HAL_I2C_Master_Receive(&hi2c1, SFM3219_ADDRESS<<1, i2c_rcv_buff, 3, 1000);
-        if (status == HAL_OK)
-        {
-            if (SF04_CheckCrc (i2c_rcv_buff, 2, i2c_rcv_buff[2]) != CHECKSUM_ERROR)
-            {
-                rawQout = (int16_t)(((uint16_t)i2c_rcv_buff[0])<<8 | i2c_rcv_buff[0]);
-            }
-            else
-            {
-                char buffer[] = I2C_CRC_ERROR_MSG;
-                HAL_UART_Transmit(&huart3, (uint8_t*)buffer, (uint8_t)(strlen(buffer)), 100U);
-                i2c_retry++;
-            }
-        }
-        else
-        {
-            char buffer[] = I2C_READ_ERROR_MSG;
-            HAL_UART_Transmit(&huart3, (uint8_t*)buffer, (uint8_t)(strlen(buffer)), 100U);
-            i2c_retry++;
-        }
-        if (i2c_retry >= 10)
-        {
-            tick_i2c = 0;
-            i2c_state = I2C_INIT;
-        }
-    }
-}
-
-
-void UpdateMeasurements()
-{
-    Measures[MEAS_QOUT]   = RawSFM3019ToLmin(rawQout, offsetQout);
-    Measures[MEAS_POUT]   = VoltageTo01mbar( RawADCToVolt( adc_buf[ADC_IN1_PA0_POUT] ), offsetPout );
-    Measures[MEAS_PPROX]  = VoltageTo01mbar( RawADCToVolt( adc_buf[ADC_IN11_PB0_PROX] ), offsetPprox );
-    Measures[MEAS_I_MOT]  = RawToCal(adc_buf[ADC_IN6_PC0_I_MOT], MAX_CURRENT, MAX_RAW_ADC);
-    Measures[MEAS_S_MOT]  = RawToCal(adc_buf[ADC_IN7_PC1_S_MOT], MAX_SPEED, MAX_RAW_ADC);
-}
-
-void PrintMeasurements()
-{
-    float tick = HAL_GetTick();
-    tick /= 1000;
-
-    float measures[6];
-    measures[0] = tick;
-    measures[1] = ((float)Measures[MEAS_QOUT])/100;
-    measures[2] = ((float)Measures[MEAS_POUT])/100;
-    measures[3] = ((float)Measures[MEAS_PPROX])/100;
-    measures[4] = ((float)Measures[MEAS_S_MOT]);
-    measures[5] = ((float)Measures[MEAS_I_MOT])/100;
-
-
-    //printFloats(measures, 6);
-    printFloatsTelePlot(measures, data_names, 6);
-}
-
 
 void Tick_1ms()
 {
-    static int time = 0;
-    static int time_ini = 0;
-    static int time_duration = 0;
-
-    ReadQoutSensor();
-    UpdateMeasurements();
-
-    if (time%10 == 0)
-    {
-        PrintMeasurements();
-    }
-
-    time += 1;
-
-    cmd_peep = 500;
-    if ( (time - time_ini) > time_duration )
-    {
-        if (cmd_motor == CMD_MOTOR_INSPI)
-        {
-            cmd_motor = 4*CMD_MOTOR_EXPI;
-            cmd_valve = 0;
-            time_duration = EXPI_TIME;
-        }
-        else
-        {
-            cmd_motor = CMD_MOTOR_INSPI;
-            cmd_valve = 999;
-            time_duration = INSPI_TIME;
-        }
-        time_ini = time;
-    }
-
-    // cmd_motor
-    // Calculate new target motor
-    // 3.3V - 4095 -> 45'000 rpm
-    //int error = target_motor_qout - Measures[MEAS_QOUT];
-    //int cmd_target_tmp = (int)cmd_motor + (error / 200);
-    //if (cmd_target_tmp < 0) cmd_target_tmp = 0;
-    //if (cmd_target_tmp > 4095) cmd_target_tmp = 4095;
-    //cmd_motor = (uint32_t)cmd_target_tmp;
-
-    DAC1->DHR12R1 = cmd_motor%4096;
-
-    UpdatePWM1(cmd_valve);
-    UpdatePWM2(cmd_peep);
-
-    UpdatePWM3(cmd_valve);
-
+    tick_main_cpp();
 }
 
 
@@ -376,6 +92,8 @@ void Tick_1ms()
   */
 int main(void)
 {
+  /* USER CODE BEGIN 1 */
+
   p_hadc1           = &hadc1;
   p_hdma_adc1       = &hdma_adc1;
   p_hdac            = &hdac;
@@ -384,8 +102,6 @@ int main(void)
   p_htim15          = &htim15;
   p_huart3          = &huart3;
   p_hdma_usart3_tx  = &hdma_usart3_tx;
-
-  /* USER CODE BEGIN 1 */
 
   /* USER CODE END 1 */
 
@@ -429,23 +145,7 @@ int main(void)
   MX_TIM2_Init();
   /* USER CODE BEGIN 2 */
 
-  main_init_cpp();
-
-  UpdatePWM1(0);
-  UpdatePWM2(0);
-  TIM15->CCER |= TIM_CCER_CC1E;
-  TIM15->CCER |= TIM_CCER_CC2E;
-  HAL_TIMEx_PWMN_Start(&htim15, HAL_TIM_ACTIVE_CHANNEL_1);
-
-  UpdatePWM3(0);
-  TIM2->CCER |= TIM_CCER_CC1E;
-  HAL_TIMEx_PWMN_Start(&htim2, HAL_TIM_ACTIVE_CHANNEL_1);
-
-  // !!! Start UART before ADC  !!! ////////
-  HAL_UART_Receive_IT(&huart3, UART3_rxBuffer, 1);
-  HAL_ADC_Start_DMA(&hadc1, (uint32_t*)adc_buf, ADC_BUF_LEN);
-
-  InitQoutSensor();
+  init_main_cpp();
 
   /* USER CODE END 2 */
 
@@ -455,7 +155,7 @@ int main(void)
   while (1)
   {
 	  HAL_Delay(10);
-	  //(void) main_cpp();
+	  loop_main_cpp();
     /* USER CODE END WHILE */
 
     /* USER CODE BEGIN 3 */
